# Vectorize ROW initialization