@DHowett 2023-06-09
Why does vectorization improve performance _printing text_ so much? I thought that all rows were eagerly initialized before your recent work.

Is this literally all from reinitializing individual rows as we recycle/circle them/