@lhecker 2023-06-09
> Is this literally all from reinitializing individual rows as we recycle/circle them?

Yes, pretty much. It reduces the row initialization cost from around 80ns per 120 columns down to 5ns. OpenConsole with all these recent changes included processes something around 1.7M rows per second, so that's why it has such a big impact. Not clearing rows at all before writing them would also be nice, for instance by maintaining the "end of the row" column, and simply leaving the rest of the row as uninitialized memory. But I believe that doing this in a robust way is a long way out whereas this was fairly easy to implement, tune and test within about an hour.